Move PageHeader to common directory
