+++ assets/js/components/HealthSummary/HomeHealthSummary.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import Table from '@common/Table';
 import HealthIcon from '@components/Health/HealthIcon';
 import { Link } from 'react-router-dom';
-import PageHeader from '@components/PageHeader';
+import PageHeader from '@common/PageHeader';
 import HealthSummary from '@components/HealthSummary';
 import useQueryStringValues from '@hooks/useQueryStringValues';
 import { getCounters, isMostRelevantPrio } from './summarySelection';
