+++ assets/js/components/HostDetails/HostSettingsPage.jsx
@@ -13,7 +13,7 @@ import { getCatalog } from '@state/selectors/catalog';
 import { getHost, getHostSelectedChecks } from '@state/selectors/host';
 import { isSaving } from '@state/selectors/checksSelection';
 
-import PageHeader from '@components/PageHeader';
+import PageHeader from '@common/PageHeader';
 import BackButton from '@common/BackButton';
 
 import ChecksSelection from '@components/ChecksSelection';
