+++ assets/js/components/HostsList.jsx
@@ -11,7 +11,7 @@ import Tags from '@common/Tags';
 import HostLink from '@components/HostLink';
 import ClusterLink from '@pages/ClusterDetails/ClusterLink';
 import SapSystemLink from '@components/SapSystemLink';
-import PageHeader from '@components/PageHeader';
+import PageHeader from '@common/PageHeader';
 import Pill from '@common/Pill';
 import HealthSummary from '@components/HealthSummary/HealthSummary';
 import { getCounters } from '@components/HealthSummary/summarySelection';
