+++ assets/js/components/SapSystemDetails/GenericSystemDetails.jsx
@@ -9,7 +9,7 @@ import { APPLICATION_TYPE, getEnsaVersionLabel } from '@lib/model/sapSystems';
 
 import ListView from '@common/ListView';
 import Table from '@common/Table';
-import PageHeader from '@components/PageHeader';
+import PageHeader from '@common/PageHeader';
 import DeregistrationModal from '@components/DeregistrationModal';
 
 import {
