+++ assets/js/pages/ChecksCatalog/ChecksCatalog.jsx
@@ -9,6 +9,7 @@ import {
   TARGET_CLUSTER,
 } from '@lib/model';
 import { ASCS_ERS, HANA_SCALE_OUT,clusterTypes, getClusterTypeLabel } from '@lib/model/clusters';
+import { clusterTypes, getClusterTypeLabel } from '@lib/model/clusters';
 import Accordion from '@common/Accordion';
 import PageHeader from '@common/PageHeader';
 import Select, { createOptionRenderer, OPTION_ALL } from '@components/Select';
